PathfindingEditorWPF: Move code to sharedpathfinding to allow experiments to work
